PI-532 added a constants module
